+++ src/module/applications/sheets/character.mjs
@@ -3,7 +3,7 @@ import { EquipmentModel, KitModel, ProjectModel } from "../../data/item/_module.
 import DrawSteelActorSheet from "./actor-sheet.mjs";
 
 /** @import { HeroTokenModel } from "../../data/settings/hero-tokens.mjs"; */
-/** @import { ActorSheetItemContext, ActorSheetEquipmentContext } from "../_types.js" */
+/** @import { ActorSheetItemContext, ActorSheetEquipmentContext } from "./_types.js" */
 
 export default class DrawSteelCharacterSheet extends DrawSteelActorSheet {
   static DEFAULT_OPTIONS = {
@@ -112,9 +112,9 @@ export default class DrawSteelCharacterSheet extends DrawSteelActorSheet {
 
   /**
    * Prepare the context for equipment categories and individual equipment items
-   * @returns {Record<keyof typeof ds["CONFIG"]["equipment"]["categories"] | "other", ActorSheetEquipmentContext>}
    */
   async _prepareEquipmentContext() {
+    /** @type {Record<string, ActorSheetEquipmentContext>} */
     const context = {};
     const equipment = this.actor.itemTypes.equipment.toSorted((a, b) => a.sort - b.sort);
 
@@ -137,6 +137,13 @@ export default class DrawSteelCharacterSheet extends DrawSteelActorSheet {
       const category = context[item.system.category] ? item.system.category : "other";
 
       context[category].equipment.push(await this._prepareItemContext(item));
+    }
+
+    // Filter out unused headers for play mode
+    if (this.isPlayMode) {
+      for (const [key, value] of Object.entries(context)) {
+        if (!value.equipment.length) delete context[key];
+      }
     }
 
     return context;